Inline user resolver & re-arrange helpers for less file jumping.
